backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
beautifulsoup4==4.14.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
cryptography==46.0.7
fastapi==0.136.0
google-ai-generativelanguage==0.6.15
google-api-core==2.30.3
google-api-python-client==2.194.0
google-auth==2.49.2
google-auth-httplib2==0.3.1
google-auth-oauthlib==1.3.1
google-generativeai==0.8.6
googleapis-common-protos==1.74.0
greenlet==3.4.0
grpcio==1.80.0
grpcio-status==1.71.2
gspread==6.2.1
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
oauthlib==3.3.1
packaging==26.1
pillow==12.2.0
playwright==1.58.0
pluggy==1.6.0
proto-plus==1.27.2
protobuf==5.29.6
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.13.2
pydantic_core==2.46.2
pyee==13.0.1
Pygments==2.20.0
pyparsing==3.3.2
pytest==9.0.3
pytest-asyncio==1.3.0
python-dotenv==1.2.2
python-multipart==0.0.26
requests==2.33.1
requests-oauthlib==2.0.0
soupsieve==2.8.3
starlette==1.0.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.44.0
